Assistant checkpoint: Fix PDF export and improve visibility control
Assistant generated file changes:
- templates/index.html: Update jsPDF initialization
- static/script.js: Update PDF export functionality and visibility

---

User prompt:

i added functionality to export the final created story as pdf, but it did not work as the text and images were not included. we tried to fix this, but now we get an error when clicking the export button: jsPDF is not defined
Can you check and fix it. I would like to make it possible, that the user exports the story with all the texts and images into a pdf. also the Export as PDF button should only appear when the story is completed as only then all images are created.